AmFFmpegExtractor: The video does not play smoothly. [1/1]
PD#OTT-9362

Problem:
Special MP4 source,video format is H264,frame rate is 15 frames,with B frames.
The order of timestamps increasing in sequence causes
confusion in the decoded timestamps.

Solution:
In this case,only the PTS of the I frame is used,
and the PTS of other frames is calculated by the amlvideo driver.

Verify:
Verify by Hisense_A311D-W400
SourceCode:
    commit 1be8e887bb2d42a22a476f78257e56f43ccaed01
    Author: yunmin.chen <yunmin.chen@amlogic.com>
    Date:   Tue Mar 31 16:00:38 2020 +0800

    AmFFmpegExtractor: The video does not play smoothly. [1/1]

    PD#OTT-9362

    Change-Id: Ifca7c0e7b0ba7c9003bd6dfa4ce740bb8a504dab
    Signed-off-by: yunmin.chen <yunmin.chen@amlogic.com>

Change-Id: Ib193e3b90eda8e07b37d2c3277c3504cb88ddeb5
Signed-off-by: yunmin.chen <yunmin.chen@amlogic.com>